Extract ApplicationInstance to separate file
